Board initalistion completed
Contains all code that sets the board up with rooms and doors, modified
the output to the console to look more like the actualy shape of the
board (using spaces). X on the board denotes the pool, to avoid
conflicts with the patio (X marks the spot of the murder).